+++ RELEASES.md
@@ -148,6 +148,7 @@ Upcoming
 - Add IBIS API docs to view all API endpoints
 - Add IAF upgrade notification when new IAF version has been released
 - Change "Show configuration status: *ALL*" to summary page (with improved performance)
+- Add active checkbox to "Upload Configuration"
 
 
 ### Non backwards compatible changes